+++ main.py
@@ -11,7 +11,6 @@ import json
 from threading import Lock
 
 import pyof.v0x01.controller2switch.common as common01
-import pyof.v0x04.controller2switch.common as common04
 from flask import jsonify, request
 from kytos.core import KytosEvent, KytosNApp, log, rest
 from kytos.core.helpers import listen_to
@@ -169,6 +168,36 @@ class GenericFlow():
                     for of_action in instruction.actions:
                         action = Action13.from_of_action(of_action)
                         flow.actions.append(action)
+        return flow
+
+    @classmethod
+    def from_replies_flows(cls, flow04):
+        """Create a flow from a flow passed on
+        replies_flows in event kytos/of_core.flow_stats.received."""
+
+        flow = GenericFlow(version='0x04')
+        flow.idle_timeout = flow04.idle_timeout.value
+        flow.hard_timeout = flow04.hard_timeout.value
+        flow.priority = flow04.priority.value
+        flow.table_id = flow04.table_id.value
+        flow.cookie = flow04.cookie.value
+        flow.duration_sec = flow04.duration_sec.value
+        flow.packet_count = flow04.packet_count.value
+        flow.byte_count = flow04.byte_count.value
+
+        for match in flow04.match.oxm_match_fields:
+            match_field = MatchFieldFactory.from_of_tlv(match)
+            field_name = match_field.name
+            if field_name == 'dl_vlan':
+                field_name = 'vlan_vid'
+            flow.match[field_name] = match_field
+        flow.actions = []
+        for instruction in flow04.instructions:
+            if instruction.instruction_type == \
+                    InstructionType.OFPIT_APPLY_ACTIONS:
+                for of_action in instruction.actions:
+                    action = Action13.from_of_action(of_action)
+                    flow.actions.append(action)
         return flow
 
     def do_match(self, args):
@@ -426,7 +455,7 @@ class Main(KytosNApp):
             'flow_id': flow_id,
             'packet_counter': flow.packet_count,
             'packet_per_second': flow.packet_count / flow.duration_sec
-            }
+        }
         return jsonify(packet_stats)
 
     @rest('bytes_count/<flow_id>')
@@ -439,7 +468,7 @@ class Main(KytosNApp):
             'flow_id': flow_id,
             'bytes_counter': flow.byte_count,
             'bits_per_second': flow.byte_count * 8 / flow.duration_sec
-            }
+        }
         return jsonify(bytes_stats)
 
     @rest('packet_count/per_flow/<dpid>')
@@ -514,24 +543,8 @@ class Main(KytosNApp):
     @listen_to('kytos/of_core.v0x01.messages.in.ofpt_stats_reply')
     def on_stats_reply_0x01(self, event):
         """Capture flow stats messages for v0x01 switches."""
-        self.handle_stats_reply_0x01(event)
-
-    def handle_stats_reply_0x01(self, event):
-        """Handle stats replies for v0x01 switches."""
         msg = event.content['message']
         if msg.body_type == common01.StatsType.OFPST_FLOW:
-            switch = event.source.switch
-            self.handle_stats_reply(msg, switch)
-
-    @listen_to('kytos/of_core.v0x04.messages.in.ofpt_multipart_reply')
-    def on_stats_reply_0x04(self, event):
-        """Capture flow stats messages for OpenFlow 1.3."""
-        self.handle_stats_reply_0x04(event)
-
-    def handle_stats_reply_0x04(self, event):
-        """Handle flow stats messages for OpenFlow 1.3."""
-        msg = event.content['message']
-        if msg.multipart_type == common04.MultipartType.OFPMP_FLOW:
             switch = event.source.switch
             self.handle_stats_reply(msg, switch)
 
@@ -565,3 +578,25 @@ class Main(KytosNApp):
             event = KytosEvent('amlight/flow_stats.flows_updated')
             event.content['switch'] = switch.dpid
             self.controller.buffers.app.put(event)
+
+    @listen_to('kytos/of_core.flow_stats.received')
+    def on_stats_received(self, event):
+        """Capture flow stats messages for OpenFlow 1.3."""
+        self.handle_stats_received(event)
+
+    def handle_stats_received(self, event):
+        """Handle flow stats messages for OpenFlow 1.3."""
+        switch = event.content['switch']
+        if 'replies_flows' in event.content:
+            replies_flows = event.content['replies_flows']
+            self.handle_stats_reply_received(switch, replies_flows)
+
+    # pylint: disable=no-self-use
+    def handle_stats_reply_received(self, switch, replies_flows):
+        """Iterate on the replies and set the generic flows"""
+        switch.generic_flows = [GenericFlow.from_replies_flows(flow)
+                                for flow in replies_flows]
+        switch.generic_flows.sort(
+                    key=lambda f: (f.priority, f.duration_sec),
+                    reverse=True
+                )