+++ tests/unit/test_main.py
@@ -17,8 +17,6 @@ from napps.kytos.of_core.v0x04.match_fields import (
 )
 from pyof.foundation.basic_types import UBInt32
 from pyof.v0x04.common.flow_instructions import InstructionType
-from pyof.v0x01.controller2switch.common import StatsType
-from pyof.v0x04.controller2switch.common import MultipartType
 
 
 # pylint: disable=too-many-public-methods, too-many-lines
@@ -301,6 +299,16 @@ class TestMain(TestCase):
         flow_stats.packet_count = 40
         return flow_stats
 
+    def _get_mocked_multipart_replies_flows(self):
+        """Helper method to create mock multipart replies flows"""
+        flow = self._get_mocked_flow_base()
+
+        instruction = MagicMock()
+        flow.instructions = [instruction]
+
+        replies_flows = [flow]
+        return replies_flows
+
     def _get_mocked_flow_base(self):
         """Helper method to create a mock flow object."""
         flow = MagicMock()
@@ -342,76 +350,47 @@ class TestMain(TestCase):
         # Check mocked flow id
         self.assertEqual(event_switch.generic_flows[0].id, 456)
 
-    @patch("napps.amlight.flow_stats.main.Main.handle_stats_reply")
-    def test_handle_stats_reply_0x01(self, mock_handle_stats):
-        """Test handle stats reply."""
-        flow_msg = MagicMock()
-        flow_msg.body = "A"
-        flow_msg.body_type = StatsType.OFPST_FLOW
-
-        switch_v0x01 = get_switch_mock("00:00:00:00:00:00:00:01", 0x01)
-
-        name = "kytos/of_core.v0x01.messages.in.ofpt_stats_reply"
-        content = {"source": switch_v0x01.connection, "message": flow_msg}
-        event = get_kytos_event_mock(name=name, content=content)
-
-        self.napp.handle_stats_reply_0x01(event)
-        mock_handle_stats.assert_called_once()
-
-    @patch("napps.amlight.flow_stats.main.Main.handle_stats_reply")
-    def test_handle_stats_reply_0x01__fail(self, mock_handle_stats):
-        """Test handle stats reply."""
-        flow_msg = MagicMock()
-        flow_msg.body = "A"
-        flow_msg.body_type = StatsType.OFPST_DESC
-
-        switch_v0x01 = get_switch_mock("00:00:00:00:00:00:00:01", 0x01)
-
-        name = "kytos/of_core.v0x01.messages.in.ofpt_stats_reply"
-        content = {"source": switch_v0x01.connection, "message": flow_msg}
-        event = get_kytos_event_mock(name=name, content=content)
-
-        self.napp.handle_stats_reply_0x01(event)
-
-        mock_handle_stats.assert_not_called()
-
-    @patch("napps.amlight.flow_stats.main.Main.handle_stats_reply")
-    def test_handle_stats_reply_0x04(self, mock_handle_stats):
-        """Test handle stats reply."""
-        flow_msg = MagicMock()
-        flow_msg.body = "A"
-        flow_msg.multipart_type = MultipartType.OFPMP_FLOW
+    @patch("napps.amlight.flow_stats.main.Main.handle_stats_reply_received")
+    def test_handle_stats_received(self, mock_handle_stats):
+        """Test handle_stats_received function."""
 
         switch_v0x04 = get_switch_mock("00:00:00:00:00:00:00:01", 0x04)
+        replies_flows = self._get_mocked_multipart_replies_flows()
+        name = "kytos/of_core.flow_stats.received"
+        content = {"switch": switch_v0x04, "replies_flows": replies_flows}
 
-        name = "kytos/of_core.v0x04.messages.in.ofpt_multipart_reply"
-        content = {"source": switch_v0x04.connection, "message": flow_msg}
         event = get_kytos_event_mock(name=name, content=content)
-        event.content["message"] = flow_msg
 
-        self.napp.handle_stats_reply_0x04(event)
-
+        self.napp.handle_stats_received(event)
         mock_handle_stats.assert_called_once()
 
-    @patch("napps.amlight.flow_stats.main.Main.handle_stats_reply")
-    def test_handle_stats_reply_0x04_fail(self, mock_handle_stats):
-        """Test handle stats reply."""
-        flow_msg = MagicMock()
-        flow_msg.body = "A"
-
-        flow_msg.multipart_type = MultipartType.OFPMP_PORT_DESC
+    @patch("napps.amlight.flow_stats.main.Main.handle_stats_reply_received")
+    def test_handle_stats_received_fail(self, mock_handle_stats):
+        """Test handle_stats_received function for
+        fail when replies_flows is not in content."""
 
         switch_v0x04 = get_switch_mock("00:00:00:00:00:00:00:01", 0x04)
-
-        name = "kytos/of_core.v0x04.messages.in.ofpt_multipart_reply"
-        content = {"source": switch_v0x04.connection, "message": flow_msg}
+        name = "kytos/of_core.flow_stats.received"
+        content = {"switch": switch_v0x04}
 
         event = get_kytos_event_mock(name=name, content=content)
-        event.content["message"] = flow_msg
 
-        self.napp.handle_stats_reply_0x04(event)
-
+        self.napp.handle_stats_received(event)
         mock_handle_stats.assert_not_called()
+
+    @patch("napps.amlight.flow_stats.main.GenericFlow.from_replies_flows")
+    def test_handle_stats_reply_received(self, mock_from_flow):
+        """Test handle_stats_reply_received call."""
+        mock_from_flow.return_value = self._get_mocked_flow_base()
+
+        self.napp.controller = MagicMock()
+
+        event_switch = MagicMock()
+        event_switch.dpid = 111
+        flows_mock = self._get_mocked_multipart_replies_flows()
+        self.napp.handle_stats_reply_received(event_switch, [flows_mock])
+
+        self.assertEqual(event_switch.generic_flows[0].id, 456)
 
 
 # pylint: disable=too-many-public-methods, too-many-lines
@@ -492,6 +471,42 @@ class TestGenericFlow(TestCase):
         flow_stats.instructions = [instruction]
 
         result = GenericFlow.from_flow_stats(flow_stats, version="0x04")
+        self.assertEqual(result.actions[0].as_dict(), action_dict)
+
+    # pylint: disable=no-member
+    def test_from_replies_flows(self):
+        """Test from_replies_flows function."""
+        replies_flow = MagicMock()
+
+        replies_flow.match.oxm_match_fields = [MatchDLVLAN(42).as_of_tlv()]
+
+        action_dict = {
+            "action_type": "output",
+            "port": UBInt32(1),
+        }
+        instruction = MagicMock()
+        instruction.instruction_type = InstructionType.OFPIT_APPLY_ACTIONS
+        instruction.actions = [Action40.from_dict(action_dict).as_of_action()]
+        replies_flow.instructions = [instruction]
+
+        result = GenericFlow.from_replies_flows(replies_flow)
+
+        self.assertEqual(result.idle_timeout, replies_flow.idle_timeout.value)
+        self.assertEqual(result.hard_timeout, replies_flow.hard_timeout.value)
+        self.assertEqual(result.priority, replies_flow.priority.value)
+        self.assertEqual(result.table_id, replies_flow.table_id.value)
+        self.assertEqual(result.cookie, replies_flow.cookie.value)
+        self.assertEqual(result.duration_sec, replies_flow.duration_sec.value)
+        self.assertEqual(result.packet_count, replies_flow.packet_count.value)
+        self.assertEqual(result.byte_count, replies_flow.byte_count.value)
+        self.assertEqual(result.duration_sec, replies_flow.duration_sec.value)
+        self.assertEqual(result.packet_count, replies_flow.packet_count.value)
+        self.assertEqual(result.byte_count, replies_flow.byte_count.value)
+
+        match_expect = MatchFieldFactory.from_of_tlv(
+            replies_flow.match.oxm_match_fields[0]
+        )
+        self.assertEqual(result.match["vlan_vid"], match_expect)
         self.assertEqual(result.actions[0].as_dict(), action_dict)
 
     def test_to_dict__x01(self):